+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Updated how errors are handled and returned to the client. All errors are now in the Problem Detail structure which contains "status", "type", "title", "detail" for errors that are returned to the client. Internally, Error objects also contain the "message" property as well as having the option to render "message" to the client, since older clients expect that value in the returned JSON.
 
+#### Added
+
+- Added new barcode sequence in the database for upcoming p-types.
+
 ### v0.7.4
 
 #### Added